Move response parsing to separate module
