+++ core/src/lib.rs
@@ -171,80 +171,7 @@ where
 }
 
 
-/// Parses a binary response into json, extracts the "result" field and tries to deserialize that
-/// to the desired type.
-fn parse_response<T>(response: &[u8], expected_id: u64) -> Result<T>
-where
-    for<'de> T: serde::Deserialize<'de>,
-{
-    let response_map = get_response_as_map(response)?;
-    let result_json = check_response_and_get_result(response_map, expected_id)?;
-    debug!("Received json result: {}", result_json);
-    serde_json::from_value::<T>(result_json).chain_err(|| {
-        ErrorKind::ResponseError("Result cannot deserialize to target type")
-    })
 }
-
-fn get_response_as_map(response: &[u8]) -> Result<serde_json::Map<String, serde_json::Value>> {
-    let response_json = serde_json::from_slice(response)
-        .chain_err(|| ErrorKind::ResponseError("Response is not valid json"))?;
-    if let serde_json::Value::Object(map) = response_json {
-        Ok(map)
-    } else {
-        Err(
-            ErrorKind::ResponseError("Response is not a json object").into(),
-        )
-    }
-}
-
-fn check_response_and_get_result(
-    mut response_map: serde_json::Map<String, serde_json::Value>,
-    expected_id: u64,
-) -> Result<serde_json::Value> {
-    ensure!(
-        response_map.remove("jsonrpc") == Some(serde_json::Value::String("2.0".to_owned())),
-        ErrorKind::ResponseError("Response is not JSON-RPC 2.0 compatible")
-    );
-    ensure!(
-        response_map.remove("id") == Some(expected_id.into()),
-        ErrorKind::ResponseError("Response id not equal to request id")
-    );
-    if let Some(error_json) = response_map.remove("error") {
-        let error = json_value_to_rpc_error(error_json)
-            .chain_err(|| ErrorKind::ResponseError("Malformed error object"))?;
-        bail!(ErrorKind::JsonRpcError(error));
-    }
-    response_map.remove("result").ok_or(
-        ErrorKind::ResponseError("Response has no \"result\" field").into(),
-    )
-}
-
-fn json_value_to_rpc_error(
-    mut error_json: serde_json::Value,
-) -> Result<jsonrpc_core::types::error::Error> {
-    let map = error_json
-        .as_object_mut()
-        .ok_or(ErrorKind::ResponseError("Error is not a json object"))?;
-    let code = map.remove("code")
-        .ok_or(ErrorKind::ResponseError("Error has no code field").into())
-        .and_then(|code| {
-            serde_json::from_value(code)
-                .chain_err(|| ErrorKind::ResponseError("Malformed code field in error"))
-        })?;
-    let message = map.get("message")
-        .and_then(|v| v.as_str())
-        .ok_or(ErrorKind::ResponseError(
-            "Error has no message field of string type",
-        ))?
-        .to_owned();
-
-    Ok(jsonrpc_core::types::error::Error {
-        code: code,
-        message: message,
-        data: map.remove("data"),
-    })
-}
-
 
 
 jsonrpc_client!(
